Trying to get 6 pages
I've moved some things around but basically everything thats significantly bigger than the rest of the things in the report need to be smaller. Also one of the tables i have taken out the explanation of the variable (like Employment : Retired) because it was taking up too much space and i feel it had already been defined in the full model. Let me know if you disagree though